requirements/base.txt
Django>=1.8
django-environ>=0.3.0
django-braces>=1.8.0
django-crispy-forms>=1.4.0
django-admin-bootstrapped>=2.4.0
django-authtools>=1.2.0
django-tinymce
easy-thumbnails>=2.2
djangorestframework
djangorestframework-yaml
djangorestframework-xml
markdown
django-filter
pytz
django-csvimport